Storing packet structure at the beginning of mbuf structure
YANFF has a limitation that parsed packets are loosen after exiting flow function.
We can't save packet structure (8 pointers to protocol headers) anywhere for the
following flow function because we need some memory allocation which is slow.

However each mbuf have "headroom" before packet payload. This "headroom" is used
for fast prepending to payload, for example while encapsulation. The length of
headroom is a property of DPDK and can be changed in dpdk/config/common_base file
in CONFIG_RTE_PKTMBUF_HEADROOM. We can use some of this memory for storing
packet structure there. For this we should:

1) Clean and prepare this headroom after receiving packet
   and after allocating mbuf for generation. This required
   removing "-g" from C compiling and using of fast AVX instructions.
2) Consider size of packet structure in prepend function
3) Change CreatePacket functions that creates packets from mbuf
   to ExtractPacket functions which extracts this memory from mbuf.
   This requires changing of prefetchs, they need to prefetch
   different line.